+++ packages/upload-api/src/lib.js
@@ -17,6 +17,7 @@ import { createService as createAdminService } from './admin.js'
 import { createService as createRateLimitService } from './rate-limit.js'
 import { createService as createUcanService } from './ucan.js'
 import { createService as createPlanService } from './plan.js'
+import { createService as createFilecoinService } from '@web3-storage/filecoin-api/storefront/service'
 
 export * from './types.js'
 
@@ -28,7 +29,10 @@ export const createServer = ({ id, codec = Legacy.inbound, ...context }) =>
     ...createRevocationChecker(context),
     id,
     codec,
-    service: createService(context),
+    service: createService({
+      ...context,
+      id
+    }),
     catch: (error) => context.errorReporter.catch(error),
   })
 
@@ -50,6 +54,7 @@ export const createService = (context) => ({
   upload: createUploadService(context),
   ucan: createUcanService(context),
   plan: createPlanService(context),
+  filecoin: createFilecoinService(context).filecoin,
 })
 
 /**